애니메이션 타이밍과 Coin 시스템 추가
애니메이션의 적절한 타이밍에 맞는 코드가 실행되도록 수정 완료

적이 죽으면 플레이어가 Coin을 얻는 기능 추가

+++ ClickerGame/Assets/Scripts/Data/Data.Contents.cs
@@ -46,7 +46,7 @@ namespace Data
         public float enemyDEF;
         public float enemyAttackSpeed;
         public float enemyRange;
-        public float enemyCoin;
+        public int enemyCoin;
     }
 
     [Serializable]
